+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/netty/NettyChannelPoolHandler.java
@@ -19,7 +19,6 @@ package io.shardingsphere.proxy.backend.netty;
 
 import io.netty.channel.Channel;
 import io.netty.channel.pool.ChannelPoolHandler;
-import io.shardingsphere.proxy.config.DataSourceConfig;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
